Split game.c into another file thread3_main.c
